Make the Wndows batch callable from any directory
The perl script NaturalDocs shuld be referred to with the same path
as the calling batch, otherwise the script would have to be in the
search path. The %~dp0 environment variable points to the directory
which the script is being executed in.

All batch arguments can bepassed to a nested batch by %*.