Add full support for JavaScript - functions and variables
JavaScript uses just two language artefacts - functions and variables.
Everything is an object and properties are a natural feature of them.
OOP in JavaScript makes use of functional objects, prototypes and
constructors, which are special use cases of those elements above.

This initial JavaScript support introduces support for the simply
declared functions and variables. Later the recognition of the OOP
structures will be added.